Integrate review comments from Matei
